Подключил js на меню и кнопку promo
