+++ gulpfile.js
@@ -19,7 +19,7 @@ gulp.task("copy", function () {
   return gulp.src([
     "source/fonts/**/*.{woff,woff2}",
     "source/img/**",
-    "source/js"
+    "source/js/**"
   ], {
     base: "source"
   })
