+++ source/catalog.html
@@ -267,7 +267,7 @@
           </div>
         </div>
       </fieldset>
-      <button class="cart-btn" type="submit">Добавить</button>
+      <button class="cart-btn cart-modal__btn" type="submit">Добавить</button>
     </form>
   </section>
   <div class="modal-overlay"></div>
