+++ source/index.html
@@ -354,9 +354,10 @@
           </div>
         </div>
       </fieldset>
-      <button class="cart-btn" type="submit">Добавить</button>
+      <button class="cart-btn cart-modal__btn" type="submit">Добавить</button>
     </form>
   </section>
   <div class="modal-overlay"></div>
+  <script src="js/script.js"></script>
 </body>
 </html>
