+++ source/sass/cart-modal.scss
@@ -33,6 +33,7 @@
   &__block {
     padding: 0;
     margin: 0;
+    margin-bottom: 30px;
     border: none;
     display: flex;
     flex-direction: column;
@@ -42,6 +43,7 @@
   &__title {
     margin-top: 0;
     margin-bottom: 16px;
+    font-family: $font-title;
     font-size: 25px;
     font-weight: 700;
     line-height: 30px;
